+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.4.0 - 2022-02-20
+## 0.5.0 - 2022-02-20
+### Changed
+- Introduce `ChunkRecordsIterator` and `IndexRecordsIterator` for
+iterating over records in the chunk and index section of rosbag file
+instead of the previous `RecordsIterator`. ([#6])
+
+[#6]: https://github.com/SkoltechRobotics/rosbag-rs/pull/6
+
+## 0.4.0 - 2022-02-19
 ### Added
 - Support for `bzip2` and `lz4` compression ([#3])